backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0

